Formatting stuff, moved yaml_utility.py to utilities
