+++ wall-e/stereo_rectification/grayscale_converter.py
@@ -1,4 +1,5 @@
 import cv2
 
+
 def convert_to_gray(img):
     return cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
